Address tests-run review feedback
